requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.2
python-dateutil==2.9.0.post0
camelot-py==0.11.0
opencv-python-headless==4.10.0.84
pypdf==4.3.1